+++ public/js/src/user/index.js
@@ -10,4 +10,26 @@ var loadUsers = function () {
     });
 };
 
-export default { loadUsers }
+
+function loadUser() {
+    $('#load').load('/user', function () {
+        $(this).find('#users-body').each(function () {
+            var that = this;
+            $.get("/users/", function (data) {
+                var html = '';
+                for (var i = 0; i < data.length; i++) {
+                    var user = data[i];
+                    html += '<tr><td><a href="#" data-id="' + user.id + '">' + (i + 1) + '</a></td><td>' + user.firstName + '</td><td>' + user.lastName + '</td>';
+                    html += '<td><a class="edit-user" href="#" data-id="' + user.id + '"><i class="material-icons">create</i></a><a class="del-user" href="#" data-id="' + user.id + '"><i aria-hidden="true" title="Delete user" class="material-icons">delete</i></a></td>' + '</tr>'
+                }
+                $(that).html(html);
+                $('#addUserModal').on('hidden.bs.modal', function (e) {
+                    loadUser();
+                })
+            });
+        });
+    });
+}
+
+
+export default { loadUsers, loadUser }
