+++ public/js/src/utils/globals.js
@@ -1,4 +1,5 @@
-const RATION = (100 / 4.805) * 6;
-const PAGE_SIZE = 10;
+export const RATION = (100 / 4.805) * 6;
+export const PAGE_SIZE = 10;
+export const UPDATE_FREQ = 1000;
 
-export default { RATION, PAGE_SIZE }
+export default { RATION, PAGE_SIZE, UPDATE_FREQ }
